requirements.txt
python-nmap==0.7.1
pyserial==3.5
pyusb==1.2.1
